:sparkles: 게임방 개설 및 입장 및 퇴장 구현

+++ server/src/main/java/com/classic/imteller/api/repository/Game.java
@@ -10,5 +10,5 @@ import javax.persistence.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Game extends BaseEntity {
     @Column(nullable = false, length=20)
-    private int session;
+    private long session;
 }
